working ADC reading, PLL stuck at MUVOUT VDD

+++ Radar_controller_example/Example_Any_Frequency/Drivers/AD7676/Inc/ad7676.h
@@ -25,7 +25,7 @@ void ad7676_init(data_Collector_TypeDef** ad7676_data);
 
 void ad7676_spi_read(uint8_t* buf, uint8_t size);
 
-float ad7676_calculate_output(uint16_t sample);
+int ad7676_calculate_output(int32_t sample);
 
 void ad7676_read_samples(uint16_t samples);
